# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateAccountRequest:
    properties:
      document_number:
        type: string
    required:
    - document_number
    type: object
  dto.CreateAccountResponse:
    properties:
      account_id:
        type: integer
      document_number:
        type: string
    type: object
  dto.CreateTransactionRequest:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      operation_type_id:
        type: integer
    type: object
  dto.CreateTransactionResponse:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      operation_type_id:
        type: integer
      transaction_id:
        type: integer
    type: object
  dto.FindAccountByIdResponse:
    properties:
      account_id:
        type: integer
      document_number:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Fábio Sartori
    url: https://github.com/kiosanim
  description: Customer Account & Transactions
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Pismo Code Assessment API
  version: "1.0"
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: Creates a new account with a valid and not used document number
      parameters:
      - description: Account Data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateAccountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an account
      tags:
      - Accounts
  /accounts/{id}:
    get:
      description: Returns an account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.FindAccountByIdResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get account by ID
      tags:
      - Accounts
  /transactions:
    post:
      consumes:
      - application/json
      description: Creates a new transaction with valid account id and document number
      parameters:
      - description: Transaction Data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a transaction
      tags:
      - Transactions
swagger: "2.0"